refactor: 폴더 이름 변경
- imageSetup/api → imageSetup/apis
