+++ src/pages/imageSetup/apis/activityInfoApi.ts
