+++ src/pages/imageSetup/apis/floorPlanApi.ts
