+++ src/pages/imageSetup/apis/houseInfoApi.ts
