+++ src/pages/imageSetup/apis/interiorStyleApi.ts
