+++ src/pages/imageSetup/hooks/useBottomSheetAddress.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { postAddress } from '../api/bottomSheetAddressApi';
+import { postAddress } from '../apis/bottomSheetAddressApi';
 import type { BottomSheetAddressRequest } from '../types/apis/bottomSheetAddress.types';
 
 export const useBottomSheetAddress = () => {
